Final refactoring - Deliver suggestion system
